chore: moved achievements
